my subjects page added with addition and removal of favourite subjects and some changes made about the login token

+++ Networking/Networking/Api Models/ApiSubjectModel.swift
@@ -10,6 +10,6 @@ public struct ApiSubjectResponse: Codable {
 }
 
 public struct ApiSubject: Codable {
-    let id: Int64?
+    let id: Int?
     let subjectName: String?
 }
